Edit container for Change Password and create popup  (#2188)
* Edit container for Change Password and create a modal popup for Password change

* Separated change password modal and used custom hook for processing modal

* Separated test files

* fixed the logic in the test files and changed the paths using alias
