+++ src/constants/translations/en/edit-profile.json
@@ -73,13 +73,15 @@
     "passwordSecurityTab": {
       "tabTitle": "Password & Security",
       "title": "Password & Security",
+      "subTitle": "You can change your password using the current one.",
       "description": "Change password and security settings",
       "changePassword": "Change password",
+      "changePasswordModalDescription": "Reset your current password to new one.",
       "changePasswordConfirm": "Are you sure you want to change the password?",
       "currentPassword": "Current password",
       "newPassword": "New password",
       "retypePassword": "Re-type new password",
-      "savePassword": "Save new password",
+      "savePassword": "Save the password",
       "deactivateAccount": "Deactivate account",
       "deactivateBtn": "Deactivate",
       "deactivateTitle": "Deactivate your account?",
