+++ src/containers/edit-profile/password-security-tab/PasswordSecurityTab.styles.ts
@@ -4,14 +4,12 @@ import { rootContainer } from '~/containers/edit-profile/common.styles'
 export const styles = {
   container: rootContainer,
   titleAndDescription: {
-    title: {
-      typography: TypographyVariantEnum.H6
-    },
-    description: {
-      typography: TypographyVariantEnum.Body2,
-      color: 'primary.500',
-      mb: '30px'
-    }
+    typography: TypographyVariantEnum.H5
+  },
+  description: {
+    typography: TypographyVariantEnum.Subtitle1,
+    color: 'primary.500',
+    mb: '30px'
   },
   subtitle: {
     typography: TypographyVariantEnum.Body1,
@@ -20,13 +18,9 @@ export const styles = {
   form: { display: 'flex', flexDirection: 'column', gap: '8px' },
 
   passwordButtonsContainer: {
-    m: '10px 0 20px',
-    display: 'grid',
-    gridTemplateColumns: {
-      sm: 'repeat(5, minmax(0, 1fr))',
-      md: 'repeat(7, minmax(0, 1fr))',
-      lg: 'repeat(10, minmax(0, 1fr))'
-    },
+    display: 'flex',
+    justifyContent: 'flex-end',
+    alignItems: 'flex-end',
     gap: '10px'
   },
 
@@ -39,5 +33,12 @@ export const styles = {
   deactivateButton: {
     mt: '20px',
     backgroundColor: 'error.700'
+  },
+  modalContainer: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    width: '700px',
+    position: 'relative'
   }
 }
