+++ src/containers/edit-profile/password-security-tab/PasswordSecurityTab.tsx
@@ -1,113 +1,26 @@
-import { useCallback, FC } from 'react'
+import { FC } from 'react'
 import { useTranslation } from 'react-i18next'
-
 import Box from '@mui/material/Box'
-import Typography from '@mui/material/Typography'
-import Divider from '@mui/material/Divider'
 
-import Loader from '~/components/loader/Loader'
-import AppButton from '~/components/app-button/AppButton'
-import AppTextField from '~/components/app-text-field/AppTextField'
-import TitleWithDescription from '~/components/title-with-description/TitleWithDescription'
-import { snackbarVariants } from '~/constants'
-
-import useForm from '~/hooks/use-form'
+import { useModalContext } from '~/context/modal-context'
 import useChangeUserStatus from '~/hooks/use-change-user-status'
-import useAxios from '~/hooks/use-axios'
-import { useAppDispatch, useAppSelector } from '~/hooks/use-redux'
-import useInputVisibility from '~/hooks/use-input-visibility'
-import useConfirm from '~/hooks/use-confirm'
-
-import { AuthService } from '~/services/auth-service'
-import { openAlert } from '~/redux/features/snackbarSlice'
+import PasswordSecurityItem from '~/containers/edit-profile/password-security-tab/password-security-item/PasswordSecurityItem'
+import ChangePasswordModal from '~/containers/edit-profile/password-security-tab/change-password-modal/ChangePasswordModal'
+import AppButton from '~/components/app-button/AppButton'
 
 import { styles } from '~/containers/edit-profile/password-security-tab/PasswordSecurityTab.styles'
-
-import {
-  ButtonVariantEnum,
-  ComponentEnum,
-  ButtonTypeEnum,
-  InputEnum,
-  SizeEnum,
-  FormValues,
-  ErrorResponse
-} from '~/types'
-import {
-  initialValues,
-  validations
-} from '~/containers/edit-profile/password-security-tab/PasswordSecurityTab.constants'
+import { ButtonVariantEnum, SizeEnum } from '~/types'
 
 const PasswordSecurityTab: FC = () => {
   const { t } = useTranslation()
-  const dispatch = useAppDispatch()
-  const { userId } = useAppSelector((state) => state.appMain)
-  const { checkConfirmation } = useConfirm()
-
+  const { openModal } = useModalContext()
   const { neededAction, checkStatusChange } = useChangeUserStatus()
 
-  const handleSubmitChangePassword = async () => {
-    const confirmed = await checkConfirmation({
-      message: t(
-        'editProfilePage.profile.passwordSecurityTab.changePasswordConfirm'
-      ),
-      title: 'titles.confirmTitle',
-      check: true
+  const openChangePasswordModal = () => {
+    openModal({
+      component: <ChangePasswordModal />
     })
-    if (confirmed) {
-      resetErrors()
-      await sendChangedPassword({
-        password: data.password,
-        currentPassword: data.currentPassword
-      })
-    }
   }
-
-  const handleResponse = () => {
-    dispatch(
-      openAlert({
-        severity: snackbarVariants.success,
-        message: 'editProfilePage.profile.successMessage'
-      })
-    )
-  }
-
-  const changePassword = useCallback(
-    (data: { password: string; currentPassword: string }) => {
-      return AuthService.changePassword(userId, data)
-    },
-    [userId]
-  )
-
-  const handleResponseError = (error?: ErrorResponse) => {
-    if (error?.code === 'INCORRECT_CREDENTIALS') {
-      handleErrors('password', t('common.errorMessages.samePasswords'))
-    } else {
-      handleErrors('currentPassword', t('common.errorMessages.currentPassword'))
-    }
-  }
-
-  const { loading, fetchData: sendChangedPassword } = useAxios({
-    service: changePassword,
-    onResponse: handleResponse,
-    onResponseError: handleResponseError,
-    fetchOnMount: false
-  })
-
-  const {
-    data,
-    handleSubmit,
-    handleInputChange,
-    errors,
-    handleBlur,
-    resetData,
-    resetErrors,
-    handleErrors
-  } = useForm<FormValues>({
-    onSubmit: handleSubmitChangePassword,
-    initialValues: initialValues,
-    validations: validations
-  })
-
   const handleChangeStatusClick = () => {
     void checkStatusChange(
       `editProfilePage.profile.passwordSecurityTab.${neededAction}Title`,
@@ -115,113 +28,27 @@ const PasswordSecurityTab: FC = () => {
       true
     )
   }
-
-  const {
-    inputVisibility: currentPasswordVisibility,
-    showInputText: showCurrentPassword
-  } = useInputVisibility(errors.currentPassword)
-
-  const { inputVisibility: passwordVisibility, showInputText: showPassword } =
-    useInputVisibility(errors.password)
-
-  const {
-    inputVisibility: newPasswordVisibility,
-    showInputText: showNewPassword
-  } = useInputVisibility(errors.confirmPassword)
-
-  const onDiscard = () => {
-    resetData()
-    resetErrors()
-  }
-
-  const saveButtonContent = loading ? (
-    <Loader size={20} />
-  ) : (
-    t('editProfilePage.profile.passwordSecurityTab.savePassword')
-  )
-
-  const inputType = (isVisible: boolean) =>
-    isVisible ? InputEnum.Text : InputEnum.Password
-
   return (
-    <Box sx={styles.container}>
-      <TitleWithDescription
-        description={t(
-          'editProfilePage.profile.passwordSecurityTab.description'
+    <Box>
+      <PasswordSecurityItem
+        buttonText={t(
+          'editProfilePage.profile.passwordSecurityTab.changePassword'
         )}
-        style={styles.titleAndDescription}
+        description={t('editProfilePage.profile.passwordSecurityTab.subTitle')}
+        onClick={openChangePasswordModal}
+        sx={styles.deactivateButton}
         title={t('editProfilePage.profile.passwordSecurityTab.title')}
       />
-      <Box component={ComponentEnum.Form} onSubmit={handleSubmit}>
-        <Typography sx={styles.subtitle}>
-          {t('editProfilePage.profile.passwordSecurityTab.changePassword')}
-        </Typography>
-        <Box sx={styles.form}>
-          <AppTextField
-            InputProps={currentPasswordVisibility}
-            errorMsg={t(errors.currentPassword)}
-            fullWidth
-            label={t(
-              'editProfilePage.profile.passwordSecurityTab.currentPassword'
-            )}
-            onBlur={handleBlur('currentPassword')}
-            onChange={handleInputChange('currentPassword')}
-            type={inputType(showCurrentPassword)}
-            value={data.currentPassword}
-          />
-          <AppTextField
-            InputProps={passwordVisibility}
-            errorMsg={t(errors.password)}
-            fullWidth
-            label={t('editProfilePage.profile.passwordSecurityTab.newPassword')}
-            onBlur={handleBlur('password')}
-            onChange={handleInputChange('password')}
-            type={inputType(showPassword)}
-            value={data.password}
-          />
-          <AppTextField
-            InputProps={newPasswordVisibility}
-            errorMsg={t(errors.confirmPassword)}
-            fullWidth
-            label={t(
-              'editProfilePage.profile.passwordSecurityTab.retypePassword'
-            )}
-            onBlur={handleBlur('confirmPassword')}
-            onChange={handleInputChange('confirmPassword')}
-            type={inputType(showNewPassword)}
-            value={data.confirmPassword}
-          />
-        </Box>
-        <Box sx={styles.passwordButtonsContainer}>
-          <AppButton
-            size={SizeEnum.Large}
-            sx={styles.saveButton}
-            type={ButtonTypeEnum.Submit}
-            variant={ButtonVariantEnum.Contained}
-          >
-            {saveButtonContent}
-          </AppButton>
-          <AppButton
-            onClick={onDiscard}
-            size={SizeEnum.Large}
-            sx={styles.discardButton}
-            variant={ButtonVariantEnum.Tonal}
-          >
-            {t('common.discard')}
-          </AppButton>
-        </Box>
-        <Divider />
-        <AppButton
-          onClick={handleChangeStatusClick}
-          size={SizeEnum.Large}
-          sx={styles.deactivateButton}
-          variant={ButtonVariantEnum.Danger}
-        >
-          {t(
-            `editProfilePage.profile.passwordSecurityTab.${neededAction}Account`
-          )}
-        </AppButton>
-      </Box>
+      <AppButton
+        onClick={handleChangeStatusClick}
+        size={SizeEnum.Large}
+        sx={styles.deactivateButton}
+        variant={ButtonVariantEnum.Danger}
+      >
+        {t(
+          `editProfilePage.profile.passwordSecurityTab.${neededAction}Account`
+        )}
+      </AppButton>
     </Box>
   )
 }
